requirements.txt
python-telegram-bot==13.15
openai>=0.27.0
apscheduler>=3.9.1